requirements.txt
Office365-REST-Python-Client>=2.6.2
redis>=5.0.0
msal>=1.34.0
requests>=2.31.0
python-dotenv>=0.21.0
mcp[cli]>=1.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
python-pptx>=0.6.23